# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface VisitAck {
  totalVisitors: number;
}

export type ResponseChoice = typeof ResponseChoice[keyof typeof ResponseChoice];


export const ResponseChoice = {
  yes: 'yes',
  maybe: 'maybe',
  not_now: 'not_now',
} as const;

export interface ResponseInput {
  response: ResponseChoice;
  /**
     * Optional proposal this response belongs to
     * @nullable
     */
  proposalId?: number | null;
}

export interface VisitorResponse {
  id: number;
  /** @nullable */
  proposalId: number | null;
  response: ResponseChoice;
  createdAt: string;
}

export type LocationType = typeof LocationType[keyof typeof LocationType];


export const LocationType = {
  coffee_shop: 'coffee_shop',
  cinema: 'cinema',
  restaurant: 'restaurant',
  beach: 'beach',
  park: 'park',
  custom: 'custom',
} as const;

export interface BookingInput {
  /** Calendar date in YYYY-MM-DD format */
  date: string;
  /** Time in HH:MM format */
  time: string;
  location: LocationType;
  /**
     * Required when location is "custom"
     * @nullable
     */
  customLocation?: string | null;
  /**
     * Optional proposal this booking belongs to
     * @nullable
     */
  proposalId?: number | null;
}

export interface Booking {
  id: number;
  /** @nullable */
  proposalId: number | null;
  date: string;
  time: string;
  location: LocationType;
  /** @nullable */
  customLocation: string | null;
  createdAt: string;
}

export interface DashboardStats {
  totalVisitors: number;
  totalBookings: number;
  acceptedCount: number;
  maybeCount: number;
  declinedCount: number;
  recentBookings: Booking[];
  recentResponses: VisitorResponse[];
}

export interface SiteSettings {
  welcomeMessage: string;
  /** @nullable */
  musicUrl: string | null;
  galleryPhotos: string[];
}

export interface SiteSettingsInput {
  /** @minLength 1 */
  welcomeMessage: string;
  /** @nullable */
  musicUrl?: string | null;
  galleryPhotos: string[];
}

export interface Proposal {
  id: number;
  slug: string;
  recipientName: string;
  /** @nullable */
  welcomeMessage: string | null;
  isActive: boolean;
  createdAt: string;
}

export interface ProposalInput {
  /** @minLength 1 */
  recipientName: string;
  /** @nullable */
  welcomeMessage?: string | null;
  /**
     * Optional custom slug; a random one is generated if omitted
     * @nullable
     */
  slug?: string | null;
}

export interface ProposalPublic {
  slug: string;
  /** @nullable */
  welcomeMessage: string | null;
  isActive: boolean;
}

export interface ProposalVerifyInput {
  /** @minLength 1 */
  name: string;
}

export interface LoginInput {
  /** @minLength 1 */
  password: string;
}

export interface AuthStatus {
  authenticated: boolean;
}

export type ListResponsesParams = {
/**
 * Filter responses to a specific proposal
 */
proposalId?: number;
};

export type ListBookingsParams = {
/**
 * Filter bookings to a specific proposal
 */
proposalId?: number;
};

export type GetStatsParams = {
/**
 * Get stats for a specific proposal (omit for global stats)
 */
proposalId?: number;
};