# lib/api-zod/src/generated/types/booking.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { LocationType } from './locationType';

export interface Booking {
  id: number;
  /** @nullable */
  proposalId: number | null;
  date: string;
  time: string;
  location: LocationType;
  /** @nullable */
  customLocation: string | null;
  createdAt: Date;
}