# lib/api-zod/src/generated/types/bookingInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { LocationType } from './locationType';

export interface BookingInput {
  /** Calendar date in YYYY-MM-DD format */
  date: string;
  /** Time in HH:MM format */
  time: string;
  location: LocationType;
  /**
     * Required when location is "custom"
     * @nullable
     */
  customLocation?: string | null;
  /**
     * Optional proposal this booking belongs to
     * @nullable
     */
  proposalId?: number | null;
}