# lib/api-zod/src/generated/types/proposalInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export interface ProposalInput {
  /** @minLength 1 */
  recipientName: string;
  /** @nullable */
  welcomeMessage?: string | null;
  /**
     * Optional custom slug; a random one is generated if omitted
     * @nullable
     */
  slug?: string | null;
}