# lib/api-zod/src/generated/types/responseChoice.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export type ResponseChoice = typeof ResponseChoice[keyof typeof ResponseChoice];


export const ResponseChoice = {
  yes: 'yes',
  maybe: 'maybe',
  not_now: 'not_now',
} as const;