# lib/api-zod/src/generated/types/siteSettings.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */

export interface SiteSettings {
  welcomeMessage: string;
  /** @nullable */
  musicUrl: string | null;
  galleryPhotos: string[];
}