# lib/api-zod/src/generated/types/visitorResponse.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.2.0
 */
import type { ResponseChoice } from './responseChoice';

export interface VisitorResponse {
  id: number;
  /** @nullable */
  proposalId: number | null;
  response: ResponseChoice;
  createdAt: Date;
}